Update Board.java
Incomplete Board with new methods checkSurroundingArea which checks the sites above, below, left and right of a site to see if the sites can still be hit (used for medium difficulty) and checkBoardForShip which checks every site on the board for sites containing a ship that hasn't been hit yet. They're both nearly working but I'm having issues with tracking / transferring the location of the sites. I'm thinking we have to assign the site a x coordinate and y coordinate when we initialize the board so I added coordinate variables and setters and getters to the site class but I cant get it to work yet. Other methods in Board will probably have to be altered slightly to track the coordinates / status, I've only tried altering initBoard so far.